June/services/june-chatterbox-tts/requirements.txt
astapi==0.111.0
uvicorn[standard]==0.30.1
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
chatterbox-tts>=0.1.4
soundfile>=0.12.1
librosa>=0.10.0
PyJWT[crypto]==2.8.0
cryptography==41.0.7
httpx==0.27.0
python-multipart==0.0.9
aiofiles>=23.0.0
numpy>=1.21.0
scipy>=1.9.0

# Resemble AI's watermark detection
perth>=0.1.0

# Optional: for audio conversion and processing
pydub>=0.25.1